update file structure with routes and controllers folders
