+++ controllers/userController.js
@@ -1,0 +1,30 @@
+exports.getAllUsers = (req, res) => {
+  console.log("xx")
+  res.status(500).json({
+    msg: "This route is not yet defined",
+  });
+};
+
+exports.createUser = (req, res) => {
+  res.status(500).json({
+    msg: "This route is not yet defined",
+  });
+};
+
+exports.getUser = (req, res) => {
+  res.status(500).json({
+    msg: "This route is not yet defined",
+  });
+};
+
+exports.updateUser = (req, res) => {
+  res.status(500).json({
+    msg: "This route is not yet defined",
+  });
+};
+
+exports.deleteUser = (req, res) => {
+  res.status(500).json({
+    msg: "This route is not yet defined",
+  });
+};
